Fixed both tests and made them pass
I think because of the first test you wrote that you want the difference for the row to default to the first value if theres only one. It seems like it should be zero because the first element would be both the high and low value for the row so the difference would be zero. It would work fine as is if we wanted it to be 0, but I added a special case to make it work for the test to pass (lines 73 & 74)

For the second test I couldn't make the test run without initializing the table so I initialized it to null and added lines 64-65 to make it pass. This seems pretty redundant though so let me know what you think. I don't think its possible for the initTableData() function to return an empty array even if the text document was only newlines because Integer.parseInt() would throw an error if the buffer string was empty, but let me know if I'm wrong.
I could have also initialized it to {{}, {}, {}} but this would have already passed because of the lines I added to pass the first test.

Let me know what you think of these edits